refactor(ia-t8): robots disallow migration + remove dead /guide/inbox/[id] forwarder
